pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["hamail <mhamail1223@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
pytest = "^8.3.2"
httpx = "^0.27.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
sqlalchemy = "^2.0.32"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
